+++ CHANGELOG.md
@@ -4,7 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2024-4-08
+## [1.1.0] - 2025-02-14
+
+### Fixed
+- Adapted to new RML-LV terminology: Iterable Field, Expression Field
+
+### Added
+- Field indexes
+
+### Changed
+- Replaced onLogicalSource by viewOn
+- Adapted handling of reference formulation and iterator in field to new spec
+
+## [1.0.0] - 2024-04-23
+
+No changes. Release linked to Zenodo for doi.
+
+## [0.0.1] - 2024-04-08
 
 ### Added
 - Mapping with old rml terms for use case Telefonica